소망가방 신청, 조회

+++ src/main/java/com/LKS/newgang/repository/SearchRepository.java
@@ -1,10 +1,13 @@
 package com.LKS.newgang.repository;
 
 import com.LKS.newgang.domain.Lecture;
+import com.LKS.newgang.domain.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface SearchRepository extends JpaRepository<Lecture, Long> {
+public interface SearchRepository extends JpaRepository<Lecture,Integer> {
     List<Lecture> findByDepartment(String departmentName);
 }
